Added The different things from Mohammed so we can push to merge
