+++ .idea/misc.xml
@@ -11,9 +11,11 @@
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/drawable/login_bg.xml" value="0.332" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/drawable/sign_in_button.xml" value="0.332" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_forgot.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_leaderboard.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_login.xml" value="0.25" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_meal_water_tracking.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_physical.xml" value="0.25" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_qeustionaire.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_siggin.xml" value="0.25" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/activity_sign_in.xml" value="0.36666666666666664" />
@@ -21,7 +23,6 @@
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/layout/toolbar.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1735" />
         <entry key="..\:/Users/17348/StudioProjects/FitLife-Fitness-Tracker/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.1735" />
-
       </map>
     </option>
   </component>
