+++ app/src/main/java/com/example/fitlife/MainActivity.java
@@ -1,0 +1,86 @@
+package com.example.fitlife;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.TextView;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+public class MainActivity extends AppCompatActivity {
+
+    Button profile, meal, phys;
+    TextView userName;
+    FirebaseAuth auth;
+    FirebaseUser user;
+    DatabaseReference reference;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        userName = findViewById(R.id.user);
+
+        auth = FirebaseAuth.getInstance();
+        user = auth.getCurrentUser();
+        String uid = user.getUid();
+        reference = FirebaseDatabase.getInstance().getReference().child("Users").child(uid);
+        reference.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                String First = snapshot.child("First Name").getValue(String.class);
+                userName.setText(First);
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+
+            }
+        });
+
+        profile = findViewById(R.id.profile);
+        profile.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(getApplicationContext(),userProfileActivity.class));
+                finish();
+            }
+        });
+
+        meal = findViewById(R.id.btnTrack);
+        meal.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(getApplicationContext(),MealWaterTracking.class));
+                finish();
+            }
+        });
+
+        phys = findViewById(R.id.physicals);
+        phys.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(getApplicationContext(),PhysicalActivity.class));
+                finish();
+            }
+        });
+
+    }
+    public void logout(View view){
+        FirebaseAuth.getInstance().signOut();//Log out of User
+        startActivity(new Intent(getApplicationContext(),LoginActivity.class));
+        finish();
+
+    }
+
+
+}
