+++ app/src/main/java/com/example/fitlife/PhysicalActivity.java
@@ -1,11 +1,26 @@
 package com.example.fitlife;
 
-import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
@@ -53,6 +68,7 @@ public class PhysicalActivity extends AppCompatActivity {
             @Override
             public void onClick(View view)
             {
+                startActivity(new Intent(getApplicationContext(),MainActivity.class));
                 finish();
             }
         });
@@ -75,6 +91,8 @@ public class PhysicalActivity extends AppCompatActivity {
 
             }
         });
+
+
 
     }
